create a separate package for testutil/network
