+++ testutil/network/util.go
@@ -1,4 +1,4 @@
-package testutil
+package network
 
 import (
 	"path/filepath"
