+++ x/bank/client/rest/tx_test.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	"github.com/cosmos/cosmos-sdk/simapp"
-	"github.com/cosmos/cosmos-sdk/testutil"
+	"github.com/cosmos/cosmos-sdk/testutil/network"
 	"github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/cosmos/cosmos-sdk/types/rest"
@@ -41,7 +41,7 @@ func (s *IntegrationTestSuite) TestCoinSend() {
 	}, stdTx.GetMsgs())
 }
 
-func submitSendReq(val *testutil.Validator, req bankrest.SendReq) (authtypes.StdTx, error) {
+func submitSendReq(val *network.Validator, req bankrest.SendReq) (authtypes.StdTx, error) {
 	url := fmt.Sprintf("%s/bank/accounts/%s/transfers", val.APIAddress, val.Address)
 
 	bz, err := val.ClientCtx.JSONMarshaler.MarshalJSON(req)
@@ -83,7 +83,7 @@ func generateSendReq(from authtypes.AccountI, amount types.Coins) bankrest.SendR
 	}
 }
 
-func getAccountInfo(val *testutil.Validator) (authtypes.AccountI, error) {
+func getAccountInfo(val *network.Validator) (authtypes.AccountI, error) {
 	url := fmt.Sprintf("%s/auth/accounts/%s", val.APIAddress, val.Address)
 
 	resp, err := rest.GetRequest(url)
